+++ htdocs/assets/js/site.js
@@ -1,1 +1,1 @@
-!function(e){var t={};function n(r){if(t[r])return t[r].exports;var o=t[r]={i:r,l:!1,exports:{}};return e[r].call(o.exports,o,o.exports,n),o.l=!0,o.exports}n.m=e,n.c=t,n.d=function(e,t,r){n.o(e,t)||Object.defineProperty(e,t,{enumerable:!0,get:r})},n.r=function(e){"undefined"!=typeof Symbol&&Symbol.toStringTag&&Object.defineProperty(e,Symbol.toStringTag,{value:"Module"}),Object.defineProperty(e,"__esModule",{value:!0})},n.t=function(e,t){if(1&t&&(e=n(e)),8&t)return e;if(4&t&&"object"==typeof e&&e&&e.__esModule)return e;var r=Object.create(null);if(n.r(r),Object.defineProperty(r,"default",{enumerable:!0,value:e}),2&t&&"string"!=typeof e)for(var o in e)n.d(r,o,function(t){return e[t]}.bind(null,o));return r},n.n=function(e){var t=e&&e.__esModule?function(){return e.default}:function(){return e};return n.d(t,"a",t),t},n.o=function(e,t){return Object.prototype.hasOwnProperty.call(e,t)},n.p="",n(n.s=0)}([function(e,t,n){e.exports=n(1)},function(e,t,n){"use strict";n.r(t);var r=function(e){return document.querySelectorAll(e)};var o=function(e,t){e.forEach(e=>{e.classList.contains(t)?e.classList.remove(t):e.classList.add(t)})};var a=function(e,t,n){let r=new XMLHttpRequest;r.open(e,t,!0),r.send(),r.onreadystatechange=function(){4===r.readyState&&r.readyState&&n(r.response)}};var c=function(e){let t=window.location.search;return new URLSearchParams(t).get(e)};var i=function(e,t=0,n=0){let r=e?e.offsetTop-n:0,o=t;window.scrollTo({left:o,top:r})};var l=function(){let e=document.getElementById("theme-switcher"),t=localStorage.getItem("theme"),n=document.getElementsByTagName("html")[0];null!==t&&(n.classList.remove("light"),n.classList.remove("dark"),n.classList.add(t),"light"===t&&(e.checked=!1),"dark"===t&&(e.checked=!0)),e.onclick=function(){this.checked?(n.classList.remove("light"),n.classList.add("dark"),localStorage.setItem("theme","dark")):(n.classList.add("light"),n.classList.remove("dark"),localStorage.setItem("theme","light"))}};let s=localStorage.getItem("lastReadItems");function d(e,t){e.forEach(e=>{if(e.intersectionRatio>0){let n=localStorage.getItem("lastReadItems"),r=e.target.dataset.timestamp;n<r&&(localStorage.setItem("lastReadItems",r),document.querySelector("#unread-items__count").innerHTML=e.target.dataset.count-1),e.target.classList.add("feed-items__item--read"),t.unobserve(e.target)}})}s||(s=0),a("GET","middleware.php?return=content&channel="+c("channel")+"&timestamp="+s,(function(e){document.getElementById("content").innerHTML=e,o(r("#unread-items"),"js-hidden"),o(r("#application-loading"),"js-hidden"),function(){let e=r(".feed-items__item"),t=r("#unread-items__count")[0].innerHTML-1,n=e[t];i(n,0,52),setTimeout(()=>{!function(e){let t=new IntersectionObserver(d,{root:null,rootMargin:"-52px 0px 0px",threshold:0});e.forEach(e=>t.observe(e))}(e)},3e3)}()})),a("GET","middleware.php?return=channels",(function(e){document.getElementById("channels").innerHTML=e})),r(".js-overlay-toggle").forEach(e=>{e.addEventListener("click",()=>{o(r(e.dataset.target),"js-hidden"),event.preventDefault()},!0)}),l()}]);
+!function(e){var t={};function n(a){if(t[a])return t[a].exports;var r=t[a]={i:a,l:!1,exports:{}};return e[a].call(r.exports,r,r.exports,n),r.l=!0,r.exports}n.m=e,n.c=t,n.d=function(e,t,a){n.o(e,t)||Object.defineProperty(e,t,{enumerable:!0,get:a})},n.r=function(e){"undefined"!=typeof Symbol&&Symbol.toStringTag&&Object.defineProperty(e,Symbol.toStringTag,{value:"Module"}),Object.defineProperty(e,"__esModule",{value:!0})},n.t=function(e,t){if(1&t&&(e=n(e)),8&t)return e;if(4&t&&"object"==typeof e&&e&&e.__esModule)return e;var a=Object.create(null);if(n.r(a),Object.defineProperty(a,"default",{enumerable:!0,value:e}),2&t&&"string"!=typeof e)for(var r in e)n.d(a,r,function(t){return e[t]}.bind(null,r));return a},n.n=function(e){var t=e&&e.__esModule?function(){return e.default}:function(){return e};return n.d(t,"a",t),t},n.o=function(e,t){return Object.prototype.hasOwnProperty.call(e,t)},n.p="",n(n.s=0)}([function(e,t,n){e.exports=n(1)},function(e,t,n){"use strict";n.r(t);var a=function(e,t,n){let a=new XMLHttpRequest;a.open(e,t,!0),a.send(),a.onreadystatechange=function(){4===a.readyState&&a.readyState&&n(a.response)}};var r=function(e){return document.querySelectorAll(e)};var o=function(){window.location.reload()};var i=function(e){let t=window.location.search;return new URLSearchParams(t).get(e)};var l=function(){let e=document.getElementById("theme-switcher"),t=localStorage.getItem("theme"),n=document.getElementsByTagName("html")[0];null!==t&&(n.classList.add(t),"light"===t&&(e.checked=!1),"dark"===t&&(e.checked=!0)),e.onclick=function(){this.checked?(n.classList.remove("light"),n.classList.add("dark"),localStorage.setItem("theme","dark")):(n.classList.add("light"),n.classList.remove("dark"),localStorage.setItem("theme","light"))}};var s=function(e,t){e.forEach(e=>{e.classList.contains(t)?e.classList.remove(t):e.classList.add(t)})};var c=function(e,t=0,n=0){let a=e?e.offsetTop-n:0,r=t;window.scrollTo({left:r,top:a})};function d(e,t){e.forEach(e=>{if(e.intersectionRatio>0){let n=localStorage.getItem("lastReadItems"),a=e.target.dataset.timestamp;n<a&&(localStorage.setItem("lastReadItems",a),unreadItemsElement.innerHTML=e.target.dataset.count-1),e.target.classList.add("feed-items__item--read"),t.unobserve(e.target)}})}var u=function(e){let t=new IntersectionObserver(d,{root:null,rootMargin:"0px",threshold:0});e.forEach(e=>t.observe(e))};var m=function(e){document.getElementById("content").innerHTML=e,window.unreadItemsElement=document.getElementById("unread-items__count");let t=r(".feed-items__item"),n=unreadItemsElement.innerHTML-1,a=t[n];c(a,0,52),s(r("#unread-items"),"js-hidden"),s(r("#application-loading"),"js-hidden"),setTimeout(()=>{u(t)},1e3)};window.lastReadTimestamps=localStorage.getItem("lastReadItems"),lastReadTimestamps||(lastReadTimestamps=0),a("GET","middleware.php?return=content&channel="+i("channel")+"&timestamp="+lastReadTimestamps,m),a("GET","middleware.php?return=channels",(function(e){document.getElementById("channels").innerHTML=e})),r(".js-overlay-toggle").forEach(e=>{e.addEventListener("click",()=>{s(r(e.dataset.target),"js-hidden"),event.preventDefault()},!0)}),r(".application-reload").forEach(e=>{e.addEventListener("click",o,!0)}),l()}]);
