+++ src/assets/js/helper/reload.js
@@ -1,0 +1,5 @@
+function reload() {
+	window.location.reload();
+}
+
+export default reload;
