+++ src/assets/js/site.js
@@ -2,89 +2,36 @@
 // # imports #
 // ###########
 
-import debounce from "./helper/debounce";
-import find from "./helper/find";
-import toggleClass from "./helper/toggleClass";
 import ajaxRequest from "./helper/ajaxRequest";
+import find from "./helper/find";
+import reload from "./helper/reload";
 import returnSearchParam from "./helper/returnSearchParam";
-import scrollToTarget from "./helper/scrollToTarget";
 import themeSwitcher from "./helper/themeSwitcher";
+import toggleClass from "./helper/toggleClass";
+
+import setupTimeline from "./tools/setupTimeline";
 
 // ###########
 // # program #
 // ###########
 
 // get latest unreadItem from saved timestamp
-let lastReadTimestamps = localStorage.getItem("lastReadItems");
+window.lastReadTimestamps = localStorage.getItem("lastReadItems");
 if (!lastReadTimestamps) {
 	lastReadTimestamps = 0;
 }
 
-// load content into the ui
+// load content and setupTimeline with response
 ajaxRequest(
 	"GET",
 	"middleware.php?return=content&channel=" +
 		returnSearchParam("channel") +
 		"&timestamp=" +
 		lastReadTimestamps,
-	afterContentLoad
+	setupTimeline
 );
 
-// do this after content has been loaded
-function afterContentLoad(response) {
-	document.getElementById("content").innerHTML = response;
-	toggleClass(find("#unread-items"), "js-hidden");
-	toggleClass(find("#application-loading"), "js-hidden");
-	scrollToFirstUnreadItem();
-}
-
-function scrollToFirstUnreadItem() {
-	// we're removing 1 since the iteration through the list starts with one
-	// this way, we can achieve zero
-	let allFeedItems = find(".feed-items__item");
-	let unreadItemsCount = find("#unread-items__count")[0].innerHTML - 1;
-	let latestItem = allFeedItems[unreadItemsCount];
-
-	// now scroll to item
-	scrollToTarget(latestItem, 0, 52);
-
-	// give it some time before executing
-	setTimeout(() => {
-		updateTimestamp(allFeedItems);
-	}, 3000);
-}
-
-// remembering the timestamp, when the item comes into view
-// IntersectionObserver Supported
-function updateTimestamp(elements) {
-	//let amount = elements.length;
-	let config = {
-		root: null,
-		rootMargin: "-52px 0px 0px",
-		threshold: 0,
-	};
-
-	let observer = new IntersectionObserver(onChange, config);
-	elements.forEach((element) => observer.observe(element));
-}
-
-function onChange(changes, observer) {
-	changes.forEach((change) => {
-		if (change.intersectionRatio > 0) {
-			let lastSavedTimestamp = localStorage.getItem("lastReadItems");
-			let currentTimestamp = change.target.dataset.timestamp;
-			if (lastSavedTimestamp < currentTimestamp) {
-				localStorage.setItem("lastReadItems", currentTimestamp);
-				document.querySelector("#unread-items__count").innerHTML =
-					change.target.dataset.count - 1;
-			}
-			change.target.classList.add("feed-items__item--read");
-			observer.unobserve(change.target);
-		}
-	});
-}
-
-// load channels into UI
+// load channels into UI-Element
 ajaxRequest("GET", "middleware.php?return=channels", injectChannels);
 function injectChannels(response) {
 	document.getElementById("channels").innerHTML = response;
@@ -100,6 +47,11 @@ find(".js-overlay-toggle").forEach((element) => {
 		},
 		true
 	);
+});
+
+// reload
+find(".application-reload").forEach((element) => {
+	element.addEventListener("click", reload, true);
 });
 
 // allow theme switching
