+++ src/assets/scss/fonts/style.css
@@ -1,19 +1,20 @@
 @font-face {
 	font-family: "icomoon";
-	src: url("../fonts/icomoon.eot?3s2q6m");
-	src: url("../fonts/icomoon.eot?3s2q6m#iefix") format("embedded-opentype"),
-		url("../fonts/icomoon.ttf?3s2q6m") format("truetype"),
-		url("../fonts/icomoon.woff?3s2q6m") format("woff"),
-		url("../fonts/icomoon.svg?3s2q6m#icomoon") format("svg");
+	src: url("../fonts/icomoon.eot?an6go2");
+	src: url("../fonts/icomoon.eot?an6go2#iefix") format("embedded-opentype"),
+		url("../fonts/icomoon.ttf?an6go2") format("truetype"),
+		url("../fonts/icomoon.woff?an6go2") format("woff"),
+		url("../fonts/icomoon.svg?an6go2#icomoon") format("svg");
 	font-weight: normal;
 	font-style: normal;
+	font-display: block;
 }
 
 [class^="icon-"],
 [class*=" icon-"] {
 	/* use !important to prevent issues with browser extensions that change fonts */
 	font-family: "icomoon" !important;
-	speak: none;
+	speak: never;
 	font-style: normal;
 	font-weight: normal;
 	font-variant: normal;
@@ -25,6 +26,9 @@
 	-moz-osx-font-smoothing: grayscale;
 }
 
+.icon-reload:before {
+	content: "\e903";
+}
 .icon-menu:before {
 	content: "\e902";
 }
